refactor preprocessor utils
