+++ src/preproc_expand.c
@@ -23,6 +23,8 @@
 #include "strbuf.h"
 #include "preproc_args.h"
 #include "preproc_builtin.h"
+#include "preproc_macro_utils.h"
+#include "preproc_paste.h"
 
 #define MAX_MACRO_DEPTH 4096
 
@@ -52,224 +54,6 @@ static int check_expand_limit(strbuf_t *sb, preproc_context_t *ctx)
     return 1;
 }
 
-/*
- * Remove trailing spaces or tabs from a string buffer.
- * Used when concatenating macro tokens with the ## operator.
- */
-static void trim_trailing_ws(strbuf_t *sb)
-{
-    while (sb->len > 0 &&
-           (sb->data[sb->len - 1] == ' ' || sb->data[sb->len - 1] == '\t')) {
-        sb->len--;
-        sb->data[sb->len] = '\0';
-    }
-}
-
-/* Parse an identifier and return its length. */
-static size_t parse_ident(const char *s)
-{
-    size_t i = 0;
-    if (!isalpha((unsigned char)s[i]) && s[i] != '_')
-        return 0;
-    i++;
-    while (isalnum((unsigned char)s[i]) || s[i] == '_')
-        i++;
-    return i;
-}
-
-/* Look up a parameter name and return the argument string if found. */
-static const char *lookup_param(const char *name, size_t len,
-                               const vector_t *params, char **args)
-{
-    for (size_t p = 0; p < params->count; p++) {
-        const char *param = ((char **)params->data)[p];
-        if (strlen(param) == len && strncmp(param, name, len) == 0)
-            return args[p];
-    }
-    return NULL;
-}
-
-/*
- * Append the stringized form of a macro parameter referenced by the
- * `#` operator starting at index `i`.  The index after the processed
- * sequence is returned.
- */
-static size_t append_stringized_param(const char *value, size_t i,
-                                      const vector_t *params, char **args,
-                                      int variadic, strbuf_t *sb)
-{
-    size_t j = i + 1;
-    while (value[j] == ' ' || value[j] == '\t')
-        j++;
-    size_t len = parse_ident(value + j);
-    if (len) {
-        const char *rep = lookup_param(value + j, len, params, args);
-        if (!rep && variadic && len == 11 &&
-            strncmp(value + j, "__VA_ARGS__", 11) == 0)
-            rep = args[params->count];
-        if (rep) {
-            strbuf_append(sb, "\"");
-            for (size_t k = 0; rep[k]; k++) {
-                char c = rep[k];
-                if (c == '\\' || c == '"')
-                    strbuf_appendf(sb, "\\%c", c);
-                else
-                    strbuf_appendf(sb, "%c", c);
-            }
-            strbuf_append(sb, "\"");
-            return j + len;
-        }
-    }
-    strbuf_append(sb, "#");
-    return i + 1;
-}
-
-/*
- * Perform the `##` token-pasting operation starting at the identifier
- * beginning at `i`.  `len` and `rep` describe this first identifier.
- * When pasting occurs `*out_i` is set to the index after the sequence
- * and non-zero is returned.
- */
-static int append_pasted_tokens(const char *value, size_t i, size_t len,
-                                const char *rep, const vector_t *params,
-                                char **args, int variadic, strbuf_t *sb,
-                                size_t *out_i)
-{
-    size_t k = i + len;
-    while (value[k] == ' ' || value[k] == '\t')
-        k++;
-    if (value[k] != '#' || value[k + 1] != '#') {
-        /* check for pasting operator before the identifier */
-        size_t j = i;
-        while (j >= 2 && (value[j-1]==' ' || value[j-1]=='\t'))
-            j--;
-        if (j >= 2 && value[j-2]=='#' && value[j-1]=='#') {
-            trim_trailing_ws(sb);
-            if (sb->len >= 2 && sb->data[sb->len-2]=='#' && sb->data[sb->len-1]=='#') {
-                sb->len -= 2;
-                sb->data[sb->len] = '\0';
-            }
-        } else {
-            return 0;
-        }
-    } else {
-        k += 2;
-        while (value[k] == ' ' || value[k] == '\t')
-            k++;
-    }
-
-    if (!value[k]) {
-        trim_trailing_ws(sb);
-        if (rep)
-            strbuf_append(sb, rep);
-        else
-            strbuf_appendf(sb, "%.*s", (int)len, value + i);
-        *out_i = k;
-    } else if (isalpha((unsigned char)value[k]) || value[k] == '_') {
-        size_t l = parse_ident(value + k);
-        const char *rep2 = lookup_param(value + k, l, params, args);
-        if (!rep2 && variadic && l == 11 &&
-            strncmp(value + k, "__VA_ARGS__", 11) == 0)
-            rep2 = args[params->count];
-        trim_trailing_ws(sb);
-        if (rep)
-            strbuf_append(sb, rep);
-        else
-            strbuf_appendf(sb, "%.*s", (int)len, value + i);
-        if (rep2)
-            strbuf_append(sb, rep2);
-        else
-            strbuf_appendf(sb, "%.*s", (int)l, value + k);
-        *out_i = k + l;
-    } else {
-        trim_trailing_ws(sb);
-        if (rep)
-            strbuf_append(sb, rep);
-        else
-            strbuf_appendf(sb, "%.*s", (int)len, value + i);
-        strbuf_appendf(sb, "%c", value[k]);
-        *out_i = k + 1;
-    }
-    return 1;
-}
-
-/*
- * Expand parameter references inside a macro body.
- * Handles substitution, stringification (#) and token pasting (##)
- * using the provided argument array.
- */
-static char *expand_params(const char *value, const vector_t *params, char **args,
-                           int variadic)
-{
-    strbuf_t sb;
-    strbuf_init(&sb);
-    /*
-     * Walk the macro body one character at a time performing the
-     * appropriate substitution or copying each literal character.
-     */
-    for (size_t i = 0; value[i];) {
-        /*
-         * Detect the `#` stringizing operator.  When a single `#` is
-         * found, convert the following macro parameter to a quoted
-         * string.
-         */
-        if (value[i] == '#' && value[i + 1] != '#') {
-            i = append_stringized_param(value, i, params, args, variadic, &sb);
-            continue;
-        }
-
-        if (value[i] == '#' && value[i + 1] == '#') {
-            strbuf_append(&sb, "##");
-            i += 2;
-            while (value[i] == ' ' || value[i] == '\t') {
-                strbuf_appendf(&sb, "%c", value[i]);
-                i++;
-            }
-            continue;
-        }
-
-        size_t len = parse_ident(value + i);
-        if (len) {
-            /* Look up the identifier as a macro parameter. */
-            const char *rep = lookup_param(value + i, len, params, args);
-            if (!rep && variadic && len == 11 &&
-                strncmp(value + i, "__VA_ARGS__", 11) == 0)
-                rep = args[params->count];
-            size_t next;
-            /*
-             * Check for the `##` token pasting operator starting at this
-             * identifier and handle concatenation when present.
-             */
-            if (append_pasted_tokens(value, i, len, rep, params, args, variadic, &sb, &next)) {
-                i = next;
-                continue;
-            }
-
-            /*
-             * If the name matched a parameter copy the replacement
-             * text, otherwise copy the identifier verbatim.
-             */
-            if (rep)
-                strbuf_append(&sb, rep);
-            else
-                strbuf_appendf(&sb, "%.*s", (int)len, value + i);
-            i += len;
-            continue;
-        }
-
-        /* No identifier or operator - emit the character as plain text */
-        strbuf_appendf(&sb, "%c", value[i]);
-        i++;
-    }
-    char *out = vc_strdup(sb.data ? sb.data : "");
-    strbuf_free(&sb);
-    if (!out)
-        vc_oom();
-    return out;
-}
-
-
-/* Expand a macro invocation and append the result to "out". */
 static int expand_macro_call(macro_t *m, char **args, vector_t *macros,
                              strbuf_t *out, int depth,
                              preproc_context_t *ctx)
